# Refactor/optimize algorithm